requirements.txt
streamlit==1.38.0
langgraph==0.2.5
langchain==0.3.0
langchain-ollama==1.0.1
python-dotenv==1.0.1
